requirements.txt
adal==1.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.0
DateTime==4.4
ez-setup==0.9
idna==3.3
msal==1.17.0
numpy==1.22.3
Office365-REST-Python-Client==2.3.11
pandas==1.4.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
zope.interface==5.4.0
